server/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.2.2
matplotlib==3.7.1
numpy==2.0.0
python-dotenv==1.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1